fix: resolve MySQL 8.0 migration errors for Docker deployment
- Fix ER_INVALID_ON_UPDATE: timestamp(3) ON UPDATE clause must include
  fractional precision (ON UPDATE CURRENT_TIMESTAMP(3)) in 0001 migration
- Fix ER_TOO_LONG_IDENT: shorten 11 constraint names exceeding MySQL's
  64-char identifier limit in 0002 legacy schema migration
- Fix ER_FK_INCOMPATIBLE_COLUMNS: change dat_fasilitas_bangunan.NO_BNG
  from smallint to int to match parent dat_op_bangunan (also in lspop.ts)
- Add drizzle.local.config.ts for running migrations from host against
  Docker-forwarded port without .env.local overriding DATABASE_URL
- Add migrate service to docker-compose.yml (profile: tools) for running
  migrations inside the Docker network

+++ lib/db/schema/lspop.ts
@@ -115,7 +115,7 @@ export const datFasilitasBangunan = mysqlTable(
   "dat_fasilitas_bangunan",
   {
     ...nopColumns(),
-    noBng: smallint("NO_BNG").notNull(),
+    noBng: int("NO_BNG").notNull(),
     kdFasilitas: char("KD_FASILITAS", { length: 2 }).notNull(),
     jmlSatuan: bigint("JML_SATUAN", { mode: "number" }).notNull().default(0),
   },
